Moving all repl functionality out of main This is to aid in the addition of command line options/other bin functions
